Migrate to IntoError
This trait connects a context selector to an error type. With this
information, we can always eagerly convert to the error type when
calling `context` / `with_context`. This has the wonderful benefit of
removing the `Context` struct and deprecating the `eager` methods;
they are now redundant!

This was driven by work to add support for extension traits for
futures, which quickly ran into type inference challenges that this
trait should help with.